requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
databases
python-dotenv
pandas
pyodbc
